Remove getHazelcastInstance() from JetService
